requirements.txt
# Core Django
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# Django Channels
channels==4.2.0
daphne==4.1.2
channels_redis==4.2.1

# Database
mysqlclient==2.2.5
psycopg2-binary==2.9.9

# Utilities
requests==2.31.0
python-dotenv==1.0.1
uvicorn==0.35.0

# Development tools
pytest==8.2.0
pytest-django==4.8.0

# CORS support
django-cors-headers==4.3.1
